pyproject.toml
[project]
name = "stacks"
version = "2.0.3"
description = "Stacks, the Terraform code pre-processor"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "cryptography>=43.0.3",
    "deepmerge>=2.0",
    "gitpython>=3.1.43",
    "jinja2>=3.1.4",
    "packaging>=24.2",
    "python-hcl2>=5.1.1",
]

[project.scripts]
stacks = "stacks:main.cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stacks"]

[tool.ruff]
line-length = 300
